requirements.txt
selenium==4.16.0
webdriver-manager==4.0.1
pandas==2.1.4
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
tqdm==4.66.1
rich==13.7.0
fake-useragent==1.4.0
nltk==3.8.1
spacy==3.7.2 